wip: move ocr core logics
