+++ core/src/main/kotlin/xyz/sevive/arcaeaoffline/core/ocr/OcrDigits.kt
