+++ core/src/main/kotlin/xyz/sevive/arcaeaoffline/core/ocr/Utils.kt
