+++ core/src/main/kotlin/xyz/sevive/arcaeaoffline/core/ocr/device/rois/definition/DeviceRoisAuto.kt
