+++ core/src/main/kotlin/xyz/sevive/arcaeaoffline/core/ocr/device/rois/masker/DeviceRoisMaskerAuto.kt
